refactor(ui): split large status and toolbar modules
